refactor: move the entire diffing logic to hooks
